更新任务处理逻辑
